Refactor: 도서관 상세 레이팅 섹션 폴더 변경
